\name{bvs-package}
\alias{bvs-package}
\alias{bvs}
\docType{package}
\title{
\packageTitle{bvs}
}
\description{
\packageDescription{bvs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bvs}
\packageIndices{bvs}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bvs}

Maintainer: \packageMaintainer{bvs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
