\name{InformBVS.NI.out}
\alias{InformBVS.NI.out}
\docType{data}

\title{Example Output From 100K iterations of sampleBVS with Informative Data}

\description{
  Output from 100K iterations of \cite{\link{sampleBVS}} with the informative study-based data set \cite{\link{InformData}}.  This was ran with inform=FALSE so that the analysis corresponds to the basic Bayesian model uncertainty framework where we assume that the effects of the predictor-level covariates are 0 (alpha=0).   
}

\usage{data(InformBVS.NI.out)}

\keyword{sample output}
