\name{Informresults.I}
\alias{Informresults.I}
\docType{data}

\title{Example Summary From 100K iterations of sampleBVS with Informative Data}

\description{
  Summary from 100K iterations of \cite{\link{sampleBVS}} with the informative study-based data set \cite{\link{InformData}} using \cite{\link{summaryBVS}}.  This was ran with inform=TRUE and gene based predictor-level covariates so that the analysis follows iBMU framework described in Quintana and Conti (submitted) where we sample that the effects of the predictor-level covariates.   
}

\usage{data(Informresults.I)}

\references{ 
  Quintana M, Conti D (Submitted). \emph{Informing Variable Selection via Bayesian Model 
  Uncertainty}.
}

\keyword{sample summary output}
