\name{Informresults.NI}
\alias{Informresults.NI}
\docType{data}

\title{Example Summary From 100K iterations of sampleBVS with Informative Data}

\description{
  Summary from 100K iterations of \cite{\link{sampleBVS}} with the informative study-based data set \cite{\link{InformData}} using \cite{\link{summaryBVS}}.  This was ran with inform=FALSE so that the analysis corresponds to the basic Bayesian model uncertainty framework where we assume that the effects of the predictor-level covariates are 0 (alpha=0).   
}

\usage{data(Informresults.NI)}


\keyword{sample summary output}
