% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_plot.R
\name{plot.summary.bvs}
\alias{plot.summary.bvs}
\title{Plot function for 'summary.bvs' object}
\usage{
\method{plot}{summary.bvs}(object, type = c("s", "r"), num_models = 100,
  num_snps = 20, num_regions = 20, plot_coef = FALSE, true_coef = NULL,
  regions = NULL, prop_cases = NULL, main = NULL, ...)
}
\arguments{
\item{object}{an object of class 'summary.bvs'}

\item{type}{specifies whether to plot the top variants ("s") or the top regions ("r")}

\item{num_models}{the number of top models to place on the x-axis}

\item{num_snps}{if type = "s", the number of the top variants to place on the y-axis}

\item{num_regions}{if type = "r", the number ofthe top regions to place on the y-axis}

\item{plot_coef}{only used for rare variant analysis when rare = TRUE and there are not multiple regions. If plot_coef = TRUE, the log(OR) of the risk index for the top models is plotted on the x-axis}

\item{true_coef}{optional vector of the true odds ratios of each of the variants to plot on the y-axis (i.e. if results are from a simulation)}

\item{regions}{optional string vector with the region name for each of the variants. By default, region names are used from the 'summary.bvs' object. Using this argument will overwrite the names in the "summary.bvs" object.}

\item{main}{optional string variable giving the title of the plot}
}
\description{
This function allows the user to create image plots of the top variants and top Regions
(any user specified set of variants such as pathways or genes) included in the top models.
Variants and Regions are ordered based on marginal BF and regional BF which are plotted on the right axis.
The width of the inclusion blocks are proportional to the posterior model probability that the variant or region is included in.
}
