\name{RareBVS.out}
\alias{RareBVS.out}
\docType{data}

\title{Example Output From 100K iterations of sampleBVS with Rare Data}

\description{
  Output from 100K iterations of \cite{\link{sampleBVS}} with the Rare variant data set \cite{\link{RareData}}.  This was ran with rare=TRUE to correspond to the BRI analysis of Quintana and Conti (2011).  
}

\usage{data(RareBVS.out)}

\references{ 
  Quintana M, Conti D (2011). \emph{Incorporating Model Uncertainty in Detecting Rare Variants: 
  The Bayesian Risk Index}. Genetic Epidemiology 35:638-649.
}


\keyword{sample output}
