% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RareData}
\alias{RareData}
\title{Rare variant data}
\format{A data frame with 1912 rows and 135 variables:}
\usage{
RareData
}
\description{
A data set that contains simulated data for rare variant analysis
}
\keyword{datasets}
