\name{RareResults}
\alias{RareResults}
\docType{data}

\title{Example Summary From 100K iterations of sampleBVS with Rare Data}

\description{
  Summary from 100K iterations of \cite{\link{sampleBVS}} with the Rare variant data set \cite{\link{RareData}} using \cite{\link{summaryBVS}}.  This was ran with rare=TRUE to correspond to the BRI analysis of Quintana and Conti (2011) and with a burnin of 1000 iterations.  
}

\usage{data(RareResults)}

\references{ 
  Quintana M, Conti D (2011). \emph{Incorporating Model Uncertainty in Detecting Rare Variants: 
  The Bayesian Risk Index}. Genetic Epidemiology 35:638-649.
}


\keyword{sample summary output}
