% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_summary.R
\name{summary.bvs}
\alias{summary.bvs}
\title{Calculates Posterior Summaries for BVS Methods}
\usage{
\method{summary}{bvs}(object, burnin = 1000, prior_cov = NULL, ...)
}
\arguments{
\item{object}{an object of class 'bvs'}

\item{burnin}{number of burn-in interations}

\item{prior_cov}{matrix of predictor-level prior covariates (required only if used in original model fit)}

\item{...}{additional arguments as required by summary S3 object}
}
\description{
This function calculates global, regional, and marginal Bayes Factors that give the strength
 of evidence of there being an association in the overall set of variants of interest,
 the individual genes of interest (if specified) and the individual variants of interest.
}
