\name{BVS-package}
\alias{BVS-package}
\alias{BVS}
\docType{package}
\title{Bayesian Variant Selection: Bayesian Model Uncertainty Techniques for Genetic Association Studies}
\description{
The functions in this package focus on analyzing case-control association studies involving a group of genetic variants.  In particular, we are interested in modeling the outcome variable as a function of a multivariate genetic profile using Bayesian model uncertainty and variable selection techniques.  The package incorporates functions to analyze data sets involving common variants as well as extensions to model rare variants via the Bayesian Risk Index (BRI of Quintana and Conti (2011)) as well as haplotypes.  Finally, the package also allows the incorporation of external biological information to inform the marginal inclusion probabilities via the iBMU (Quintana and Conti (submitted)). 
}

\details{
\tabular{ll}{
Package: \tab BVS\cr
Version: \tab 4.12.0\cr
Date: \tab 2012-4-17\cr
Depends: \tab MASS, msm, haplo.stats\cr
License: \tab GPL-2\cr
}
}

\author{Melanie Quintana <maw27.wilson@gmail.com>}

\references{
   Quintana M, Conti D (2011). \emph{Incorporating Model Uncertainty in Detecting Rare Variants: 
  The Bayesian Risk Index}. Genetic Epidemiology 35:638-649.

  Quintana M, Conti D (Submitted). \emph{Integrative Variable Selection via Bayesian Model 
  Uncertainty}.
}

\keyword{package}

