% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs.R
\name{bvs}
\alias{bvs}
\title{Bayesian Variable Selection}
\usage{
bvs(y, x, forced = NULL, intercept = TRUE, family = c("gaussian",
  "binomial"), method = c("enumerate", "sample"), rare = FALSE,
  regions = NULL, prior_cov = NULL, a1 = 0, hap = FALSE, iter = 10000,
  control = list())
}
\arguments{
\item{y}{outcome variable}

\item{x}{predictor design matrix}

\item{forced}{(optional) \eqn{n x c} matrix of \emph{c} confounding variables that one wishes to adjust the
analysis for and that will be forced into every model.}

\item{intercept}{indicates whether model should include an intercept in all models. Default is TRUE.}

\item{family}{specifies error distribution for outcome variable, options include
\itemize{
   \item "gaussian"
   \item "binomial"
}}

\item{method}{specifies how to search the model space
\itemize{
   \item "enumerate": computes and summarizes all possible models in model space. Not recommended for problems where \eqn{p > 20}.
   \item "sample": performs basic Metropolis-Hastings algorithm to sample models from model space. For informative
   marginal inclusion probabilities, the algorithm also performs basic MCMC algorithm to sample the effects of predictor-level
   covariates (alpha).
}}

\item{rare}{if rare = TRUE, corresponds to the Bayesian Risk Index (BRI) algorithm of Quintana and Conti (2011) that constructs
a risk index based on the multiple rare variants within each model. The marginal likelihood of each model is then calculated
based on the corresponding risk index.}

\item{regions}{(optional) \eqn{p x 1} character or factor vector that identifies a user-defined region for each variant. If
rare = TRUE, then multiple region-specific risk indices are computed for each model.}

\item{prior_cov}{(optional) if method = "sample", a \eqn{p x q} matrix of \emph{q} predictor-level covariates that the user
wishes to incorporate into the estimation of the marginal inclusion probabilities using the iBMU algorithm.}

\item{a1}{(optional) if method = "enumerate", a \eqn{q x 1} vector of specified effects of each predictor-level covariate.}

\item{hap}{(not yet implemented) if hap = TRUE, esimtate a set of haplotypes from the multiple variants within each moel and the marginal likelihood
of each model is calculated based on the set of haplotypes.}

\item{iter}{if method = "sample", the number of iterations to run the algorithm.}

\item{control}{specifies 'bvs' control object.}
}
\description{
Bayesian Variable Selection
}
