\name{RareData}
\alias{RareData}
\docType{data}
\title{Simulated Example Rare Variant data set.}

\description{
  Simulated data set of 134 rare variants.  The
  first column represents the disease status of the individual, the
  remaining columns the counts of minor alleles (0|1|2) for each variant.}

\usage{data(RareData)}

\format{
  A data frame with 1912 observations on the following 135 variables
(case, rare variants 1:134).
}


\keyword{datasets}
